requirements.txt
# Core LLM and LangGraph
langgraph>=0.0.30
langchain>=0.1.16
langchain-openai>=0.1.3
langchain-xai>=0.0.5
langchain_community>=0.1.0

# Streamlit frontend
streamlit>=1.32.2

# PDF, Web and Image Parsing
pymupdf>=1.23.19         # PDF parsing
easyocr>=1.7.1      # OCR for images
pillow>=10.2.0            # Image handling
requests>=2.31.0          # For web scraping
beautifulsoup4>=4.12.3    # HTML parser for scraping

# Web Search
googlesearch-python>=1.2.3  # For fallback search

# Embeddings and Vector DB
sentence-transformers>=2.6.1
chromadb>=0.4.24

# Dotenv for environment variables
python-dotenv>=1.0.1

# Optional LLM support (Open Source fallback)
transformers>=4.39.3
torch>=2.2.2
accelerate>=0.28.0